Add folders in package directory
