+++ rgbd_mocap/markers/marker.py
